ディレクトリ分け、build内のref.read修正
