+++ lib/qiita/model/api/qiita_api_client.dart
@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart';
-import 'package:flutter_tutorial/qiita/model/qiita_state.dart';
+import 'package:flutter_tutorial/qiita/model/qiita_item.dart';
 import 'package:retrofit/retrofit.dart';
 
 part 'qiita_api_client.g.dart';
@@ -8,9 +8,8 @@ part 'qiita_api_client.g.dart';
 abstract class QiitaApiClient {
   factory QiitaApiClient(Dio dio, {String baseUrl}) = _QiitaApiClient;
 
-  @GET('/tags/{tag}/items/')
+  @GET('/tags/{tag}/items')
   Future<List<QiitaItem>> fetchQiitaItems(
     @Path('tag') String tag,
-    @Query('per_page') int perPage,
   );
 }
