+++ lib/qiita/model/qiita_state.freezed.dart
@@ -210,350 +210,3 @@ abstract class _QiitaState implements QiitaState {
   _$$QiitaStateImplCopyWith<_$QiitaStateImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
-
-QiitaItem _$QiitaItemFromJson(Map<String, dynamic> json) {
-  return _QiitaItem.fromJson(json);
-}
-
-/// @nodoc
-mixin _$QiitaItem {
-  @JsonKey(name: 'title')
-  String? get title => throw _privateConstructorUsedError;
-  @JsonKey(name: 'url')
-  String? get url => throw _privateConstructorUsedError;
-  @JsonKey(name: 'user')
-  QiitaUser? get user => throw _privateConstructorUsedError;
-
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
-  $QiitaItemCopyWith<QiitaItem> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $QiitaItemCopyWith<$Res> {
-  factory $QiitaItemCopyWith(QiitaItem value, $Res Function(QiitaItem) then) =
-      _$QiitaItemCopyWithImpl<$Res, QiitaItem>;
-  @useResult
-  $Res call(
-      {@JsonKey(name: 'title') String? title,
-      @JsonKey(name: 'url') String? url,
-      @JsonKey(name: 'user') QiitaUser? user});
-
-  $QiitaUserCopyWith<$Res>? get user;
-}
-
-/// @nodoc
-class _$QiitaItemCopyWithImpl<$Res, $Val extends QiitaItem>
-    implements $QiitaItemCopyWith<$Res> {
-  _$QiitaItemCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? title = freezed,
-    Object? url = freezed,
-    Object? user = freezed,
-  }) {
-    return _then(_value.copyWith(
-      title: freezed == title
-          ? _value.title
-          : title // ignore: cast_nullable_to_non_nullable
-              as String?,
-      url: freezed == url
-          ? _value.url
-          : url // ignore: cast_nullable_to_non_nullable
-              as String?,
-      user: freezed == user
-          ? _value.user
-          : user // ignore: cast_nullable_to_non_nullable
-              as QiitaUser?,
-    ) as $Val);
-  }
-
-  @override
-  @pragma('vm:prefer-inline')
-  $QiitaUserCopyWith<$Res>? get user {
-    if (_value.user == null) {
-      return null;
-    }
-
-    return $QiitaUserCopyWith<$Res>(_value.user!, (value) {
-      return _then(_value.copyWith(user: value) as $Val);
-    });
-  }
-}
-
-/// @nodoc
-abstract class _$$QiitaItemImplCopyWith<$Res>
-    implements $QiitaItemCopyWith<$Res> {
-  factory _$$QiitaItemImplCopyWith(
-          _$QiitaItemImpl value, $Res Function(_$QiitaItemImpl) then) =
-      __$$QiitaItemImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call(
-      {@JsonKey(name: 'title') String? title,
-      @JsonKey(name: 'url') String? url,
-      @JsonKey(name: 'user') QiitaUser? user});
-
-  @override
-  $QiitaUserCopyWith<$Res>? get user;
-}
-
-/// @nodoc
-class __$$QiitaItemImplCopyWithImpl<$Res>
-    extends _$QiitaItemCopyWithImpl<$Res, _$QiitaItemImpl>
-    implements _$$QiitaItemImplCopyWith<$Res> {
-  __$$QiitaItemImplCopyWithImpl(
-      _$QiitaItemImpl _value, $Res Function(_$QiitaItemImpl) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? title = freezed,
-    Object? url = freezed,
-    Object? user = freezed,
-  }) {
-    return _then(_$QiitaItemImpl(
-      title: freezed == title
-          ? _value.title
-          : title // ignore: cast_nullable_to_non_nullable
-              as String?,
-      url: freezed == url
-          ? _value.url
-          : url // ignore: cast_nullable_to_non_nullable
-              as String?,
-      user: freezed == user
-          ? _value.user
-          : user // ignore: cast_nullable_to_non_nullable
-              as QiitaUser?,
-    ));
-  }
-}
-
-/// @nodoc
-@JsonSerializable()
-class _$QiitaItemImpl implements _QiitaItem {
-  const _$QiitaItemImpl(
-      {@JsonKey(name: 'title') this.title,
-      @JsonKey(name: 'url') this.url,
-      @JsonKey(name: 'user') this.user});
-
-  factory _$QiitaItemImpl.fromJson(Map<String, dynamic> json) =>
-      _$$QiitaItemImplFromJson(json);
-
-  @override
-  @JsonKey(name: 'title')
-  final String? title;
-  @override
-  @JsonKey(name: 'url')
-  final String? url;
-  @override
-  @JsonKey(name: 'user')
-  final QiitaUser? user;
-
-  @override
-  String toString() {
-    return 'QiitaItem(title: $title, url: $url, user: $user)';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$QiitaItemImpl &&
-            (identical(other.title, title) || other.title == title) &&
-            (identical(other.url, url) || other.url == url) &&
-            (identical(other.user, user) || other.user == user));
-  }
-
-  @JsonKey(ignore: true)
-  @override
-  int get hashCode => Object.hash(runtimeType, title, url, user);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$QiitaItemImplCopyWith<_$QiitaItemImpl> get copyWith =>
-      __$$QiitaItemImplCopyWithImpl<_$QiitaItemImpl>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$QiitaItemImplToJson(
-      this,
-    );
-  }
-}
-
-abstract class _QiitaItem implements QiitaItem {
-  const factory _QiitaItem(
-      {@JsonKey(name: 'title') final String? title,
-      @JsonKey(name: 'url') final String? url,
-      @JsonKey(name: 'user') final QiitaUser? user}) = _$QiitaItemImpl;
-
-  factory _QiitaItem.fromJson(Map<String, dynamic> json) =
-      _$QiitaItemImpl.fromJson;
-
-  @override
-  @JsonKey(name: 'title')
-  String? get title;
-  @override
-  @JsonKey(name: 'url')
-  String? get url;
-  @override
-  @JsonKey(name: 'user')
-  QiitaUser? get user;
-  @override
-  @JsonKey(ignore: true)
-  _$$QiitaItemImplCopyWith<_$QiitaItemImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-QiitaUser _$QiitaUserFromJson(Map<String, dynamic> json) {
-  return _QiitaUser.fromJson(json);
-}
-
-/// @nodoc
-mixin _$QiitaUser {
-  @JsonKey(name: 'profile_image_url')
-  String? get profileImageUrl => throw _privateConstructorUsedError;
-
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
-  $QiitaUserCopyWith<QiitaUser> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $QiitaUserCopyWith<$Res> {
-  factory $QiitaUserCopyWith(QiitaUser value, $Res Function(QiitaUser) then) =
-      _$QiitaUserCopyWithImpl<$Res, QiitaUser>;
-  @useResult
-  $Res call({@JsonKey(name: 'profile_image_url') String? profileImageUrl});
-}
-
-/// @nodoc
-class _$QiitaUserCopyWithImpl<$Res, $Val extends QiitaUser>
-    implements $QiitaUserCopyWith<$Res> {
-  _$QiitaUserCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? profileImageUrl = freezed,
-  }) {
-    return _then(_value.copyWith(
-      profileImageUrl: freezed == profileImageUrl
-          ? _value.profileImageUrl
-          : profileImageUrl // ignore: cast_nullable_to_non_nullable
-              as String?,
-    ) as $Val);
-  }
-}
-
-/// @nodoc
-abstract class _$$QiitaUserImplCopyWith<$Res>
-    implements $QiitaUserCopyWith<$Res> {
-  factory _$$QiitaUserImplCopyWith(
-          _$QiitaUserImpl value, $Res Function(_$QiitaUserImpl) then) =
-      __$$QiitaUserImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({@JsonKey(name: 'profile_image_url') String? profileImageUrl});
-}
-
-/// @nodoc
-class __$$QiitaUserImplCopyWithImpl<$Res>
-    extends _$QiitaUserCopyWithImpl<$Res, _$QiitaUserImpl>
-    implements _$$QiitaUserImplCopyWith<$Res> {
-  __$$QiitaUserImplCopyWithImpl(
-      _$QiitaUserImpl _value, $Res Function(_$QiitaUserImpl) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? profileImageUrl = freezed,
-  }) {
-    return _then(_$QiitaUserImpl(
-      profileImageUrl: freezed == profileImageUrl
-          ? _value.profileImageUrl
-          : profileImageUrl // ignore: cast_nullable_to_non_nullable
-              as String?,
-    ));
-  }
-}
-
-/// @nodoc
-@JsonSerializable()
-class _$QiitaUserImpl implements _QiitaUser {
-  const _$QiitaUserImpl(
-      {@JsonKey(name: 'profile_image_url') this.profileImageUrl});
-
-  factory _$QiitaUserImpl.fromJson(Map<String, dynamic> json) =>
-      _$$QiitaUserImplFromJson(json);
-
-  @override
-  @JsonKey(name: 'profile_image_url')
-  final String? profileImageUrl;
-
-  @override
-  String toString() {
-    return 'QiitaUser(profileImageUrl: $profileImageUrl)';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$QiitaUserImpl &&
-            (identical(other.profileImageUrl, profileImageUrl) ||
-                other.profileImageUrl == profileImageUrl));
-  }
-
-  @JsonKey(ignore: true)
-  @override
-  int get hashCode => Object.hash(runtimeType, profileImageUrl);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$QiitaUserImplCopyWith<_$QiitaUserImpl> get copyWith =>
-      __$$QiitaUserImplCopyWithImpl<_$QiitaUserImpl>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$QiitaUserImplToJson(
-      this,
-    );
-  }
-}
-
-abstract class _QiitaUser implements QiitaUser {
-  const factory _QiitaUser(
-          {@JsonKey(name: 'profile_image_url') final String? profileImageUrl}) =
-      _$QiitaUserImpl;
-
-  factory _QiitaUser.fromJson(Map<String, dynamic> json) =
-      _$QiitaUserImpl.fromJson;
-
-  @override
-  @JsonKey(name: 'profile_image_url')
-  String? get profileImageUrl;
-  @override
-  @JsonKey(ignore: true)
-  _$$QiitaUserImplCopyWith<_$QiitaUserImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
