+++ lib/qiita/view_model/qiita_view_model.dart
@@ -2,28 +2,33 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_tutorial/qiita/model/qiita_state.dart';
 import 'package:flutter_tutorial/qiita/repository/qiita_repository.dart';
 
+// QiitaViewModel のインスタンスを生成
 final qiitaViewModelProvider =
     StateNotifierProvider.autoDispose<QiitaViewModel, QiitaState>(
   QiitaViewModel.new,
 );
 
+// QiitaState の状態を管理
 class QiitaViewModel extends StateNotifier<QiitaState> {
   QiitaViewModel(this._ref) : super(const QiitaState());
 
   final Ref _ref;
 
+  // QiitaRepositoryの取得
   QiitaRepository get _qiitaRepository => _ref.read(qiitaRepositoryProvider);
 
-  // 取得記事初期値設定しておく
-  Future<void> fetchQiitaItems(String tag, {int perPage = 10}) async {
+  // Qiitaアイテムの取得
+  Future<void> fetchQiitaItems(String tag) async {
     state = state.copyWith(isLoading: true);
 
-    final qiitaItems = await _qiitaRepository.fetchQiitaItems(tag, perPage);
+    // リポジトリからデータ取得
+    final qiitaItems = await _qiitaRepository.fetchQiitaItems(tag);
 
     if (!mounted) {
       return;
     }
 
+    // 状態を更新
     if (qiitaItems.isNotEmpty) {
       state = state.copyWith(
         isLoading: false,
@@ -40,6 +45,7 @@ class QiitaViewModel extends StateNotifier<QiitaState> {
     }
   }
 
+  // ホームに戻る処理
   Future<void> onBackHome() async {
     state = state.copyWith(
       isLoading: false,
